cleaned up navbar - created new components for each element
