+++ src/components/NavBar/NavBar.styles.jsx
@@ -1,6 +1,5 @@
 import React from "react";
 import styled from "styled-components";
-import { Link } from "react-router-dom";
 
 export const StyledNav = styled.nav`
   max-width: 1720px;
@@ -13,154 +12,22 @@ export const StyledNav = styled.nav`
   position: relative;
   z-index: 5;
 `;
-export const StyledLink = styled(Link)`
-  text-decoration: none;
-  color: ${(props) => props.theme.text};
-  font-size: 20px;
-  z-index: 10;
-`;
-export const StyledList = styled.ul`
+export const LinksList = styled.ul`
   display: flex;
   z-index: 10;
   width: 50%;
 `;
-export const ListItemCoins = styled.li`
-  list-style-type: none;
-  width: 100px;
-  height: 53px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin: 10px 40px 10px 10px;
-  border-radius: 10px;
-  z-index: 10;
-  background: {
-    this.props.state.coins: "blue" ? "green";
-  }
-`;
-export const ListItemPortfolio = styled.li`
-  list-style-type: none;
-  width: 100px;
-  height: 53px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin: 10px;
-  border-radius: 10px;
-  z-index: 10;
-`;
-export const NavInnerContainer = styled.div`
+export const NavBarInnerContainer = styled.div`
   width: 50%;
   display: flex;
   align-items: center;
   justify-content: space-between;
 `;
-export const SearchWrapper = styled.div`
-  width: 70%;
-`;
-const SearchDiv = styled.div`
-  z-index: 10;
-`;
-const SearchInput = styled.input`
-  width: 410px;
-  height: 53px;
-  background: ${(props) => props.theme.background};
-  color: ${(props) => props.theme.text};
-  padding-left: 80px;
-  font-size: 20px;
-  border: none;
-  border-radius: 10px;
-  z-index: 10;
-`;
-const StyledSearchSvg = styled.svg`
-  width: 30px;
-  position: absolute;
-  z-index: 10;
-  margin-left: 20px;
-  margin-top: 12px;
-  z-index: 10;
-`;
-const SearchSvg = () => {
-  return (
-    <StyledSearchSvg
-      xmlns="http://www.w3.org/2000/svg"
-      fill="none"
-      viewBox="0 0 24 24"
-      strokeWidth={1.5}
-      stroke="currentColor"
-      className="w-6 h-6"
-    >
-      <path
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        d="M21 21l-5.197-5.197m0 0A7.5 7.5 0 105.196 5.196a7.5 7.5 0 0010.607 10.607z"
-      />
-    </StyledSearchSvg>
-  );
-};
-export const Search = (props) => {
-  return (
-    <SearchDiv>
-      <SearchSvg />
-      <SearchInput placeholder="Search..." />
-    </SearchDiv>
-  );
-};
 export const CurrencyDiv = styled.div`
   z-index: 10;
   display: flex;
   align-items: center;
 `;
-export const CurrencySelect = styled.select`
-  width: 130px;
-  height: 43px;
-  background: ${(props) => props.theme.background};
-  color: ${(props) => props.theme.text};
-  text-align: center;
-  font-size: 20px;
-  padding-left: 20px;
-  border-radius: 10px;
-  z-index: 10;
-`;
-export const Option = styled.option`
-  padding-left: 15px;
-`;
-const ThemeSwitchDiv = styled.div`
-  width: 43px;
-  height: 43px;
-  background: ${(props) => props.theme.background};
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin: 10px 10px 10px 10px;
-  border-radius: 10px;
-  z-index: 10;
-`;
-const StyledMoonSvg = styled.svg`
-  width: 28px;
-  fill: ${(props) => props.theme.moonIconFill};
-  border: none;
-  z-index: 10;
-`;
-export const ThemeSwitch = (props) => {
-  return (
-    <ThemeSwitchDiv onClick={props.onClick}>
-      <StyledMoonSvg
-        xmlns="http://www.w3.org/2000/svg"
-        viewBox="0 0 24 24"
-        strokeWidth={1.5}
-        stroke="currentColor"
-        className="w-6 h-6"
-      >
-        <path
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          d="M21.752 15.002A9.718 9.718 0 0118 15.75c-5.385 0-9.75-4.365-9.75-9.75 0-1.33.266-2.597.748-3.752A9.753 9.753 0 003 11.25C3 16.635 7.365 21 12.75 21a9.753 9.753 0 009.002-5.998z"
-        />
-      </StyledMoonSvg>
-    </ThemeSwitchDiv>
-  );
-};
 export const SubNavContainer = styled.div`
   width: 100%;
   height: 43px;
@@ -193,17 +60,6 @@ export const StyledListSubNav = styled.ul`
   align-items: center;
   font-size: 16px;
 `;
-export const StyledListItemSubNav = styled.li`
-  list-style-type: none;
-  cursor: pointer;
-`;
-export const ListItemWithProgress = styled.li`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  list-style-type: disc;
-  cursor: pointer;
-`;
 export const BulletPoint = styled.div`
   width: 10px;
   height: 10px;
@@ -211,18 +67,16 @@ export const BulletPoint = styled.div`
   background: ${(props) => props.theme.bulletPointBackground};
   margin-right: 10px;
 `;
-
-export const ArrowTrendingDown = styled.svg`
+const ArrowTrendingDown = styled.svg`
   width: 25px;
   color: red;
   margin-left: 5px;
 `;
-export const ArrowTrendingUp = styled.svg`
+const ArrowTrendingUp = styled.svg`
   width: 25px;
   color: green;
   margin-left: 5px;
 `;
-
 export const TrendingDown = () => {
   return (
     <ArrowTrendingDown
@@ -241,7 +95,6 @@ export const TrendingDown = () => {
     </ArrowTrendingDown>
   );
 };
-
 export const TrendingUp = () => {
   return (
     <ArrowTrendingUp
