+++ src/components/NavBarPortfolioLink/index.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { ListItemPortfolio, StyledLink } from "./NavBarPortfolioLink.styles";
+
+export const NavBarPortfolioLink = () => {
+  return (
+    <ListItemPortfolio>
+      <StyledLink to="/portfolio">Portfolio</StyledLink>
+    </ListItemPortfolio>
+  );
+};
