+++ src/components/NavBarSearch/index.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { SearchWrapper, Search } from "./NavBarSearch.styles";
+
+export const NavBarSearch = () => {
+  return (
+    <SearchWrapper>
+      <form>
+        <Search />
+      </form>
+    </SearchWrapper>
+  );
+};
