+++ src/components/ThemeSwitch/index.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+import { ThemeSwitchStyled } from "./ThemeSwitch.styles";
+
+export const ThemeSwitch = ({ onClick }) => {
+  return <ThemeSwitchStyled onClick={onClick} />;
+};
